+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/InputComponent.java
@@ -1,8 +1,12 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
+import org.hamcrest.Matchers;
+
 import at.porscheinformatik.seleniumcomponents.AbstractSeleniumComponent;
 import at.porscheinformatik.seleniumcomponents.EditableSeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.SeleniumAsserts;
 import at.porscheinformatik.seleniumcomponents.SeleniumComponent;
+import at.porscheinformatik.seleniumcomponents.Utils;
 import at.porscheinformatik.seleniumcomponents.WebElementSelector;
 
 /**
@@ -34,13 +38,30 @@ public class InputComponent extends AbstractSeleniumComponent implements Editabl
     }
 
     /**
-     * Clears the input field an types all specified values.
+     * Clears the input field an types all specified values. Checks, that the input field finally contains the value. If
+     * the input field has a special behavior (e.g. auto typing or character restrictions) use the
+     * {@link #type(CharSequence...)} method instead.
      *
      * @param values one or more values to enter
      */
     public void enter(CharSequence... values)
     {
         clear();
+        sendKeys(values);
+
+        String jointValues = String.join("", values);
+
+        SeleniumAsserts.assertThatSoon(String.format("Enter \"%s\"", jointValues), () -> Utils.simplify(getValue()),
+            Matchers.is(Utils.simplify(jointValues)));
+    }
+
+    /**
+     * Types the value into the input field. Does not verify, that the field contains the value after typing.
+     *
+     * @param values one or more values to enter
+     */
+    public void type(CharSequence... values)
+    {
         sendKeys(values);
     }